+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/usage/port/in/command/UseReservationCommand.java
@@ -1,4 +1,4 @@
-package com.seatliberator.seatliberator.reservation.application.verification.port.in.command;
+package com.seatliberator.seatliberator.reservation.application.usage.port.in.command;
 
 import com.seatliberator.seatliberator.identity.core.actor.Actor;
