+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/usage/service/ReservationUsageService.java
@@ -1,12 +1,12 @@
-package com.seatliberator.seatliberator.reservation.application.verification.service;
+package com.seatliberator.seatliberator.reservation.application.usage.service;
 
 import com.seatliberator.seatliberator.reservation.application.booking.contract.ReservationOwnershipPolicy;
 import com.seatliberator.seatliberator.reservation.application.booking.port.out.ReservationReader;
 import com.seatliberator.seatliberator.reservation.application.shared.exception.ReservationApplicationErrorCode;
 import com.seatliberator.seatliberator.reservation.application.shared.exception.ReservationApplicationException;
-import com.seatliberator.seatliberator.reservation.application.verification.port.in.UseReservationUseCase;
-import com.seatliberator.seatliberator.reservation.application.verification.port.in.command.UseReservationCommand;
-import com.seatliberator.seatliberator.reservation.application.verification.port.in.result.UseReservationResult;
+import com.seatliberator.seatliberator.reservation.application.usage.port.in.UseReservationUseCase;
+import com.seatliberator.seatliberator.reservation.application.usage.port.in.command.UseReservationCommand;
+import com.seatliberator.seatliberator.reservation.application.usage.port.in.result.UseReservationResult;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -14,7 +14,7 @@ import java.time.Clock;
 
 @Service
 @RequiredArgsConstructor
-public class VerificationService implements
+public class ReservationUsageService implements
         UseReservationUseCase {
     private final ReservationReader reader;
     private final ReservationOwnershipPolicy ownershipPolicy;
