--- reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/verification/port/in/UseReservationUseCase.java
@@ -1,8 +1,0 @@
-package com.seatliberator.seatliberator.reservation.application.verification.port.in;
-
-import com.seatliberator.seatliberator.reservation.application.verification.port.in.command.UseReservationCommand;
-import com.seatliberator.seatliberator.reservation.application.verification.port.in.result.UseReservationResult;
-
-public interface UseReservationUseCase {
-    UseReservationResult use(UseReservationCommand command);
-}
